Update profile.php
It added comments and functionality to the page. The page currently has a user container with an image of the user. It has User data that populates from the database. It has a nav bar that views, edits and adds. If you click on any of them, it will take you to another nav bar that allows users to view, add, or edit depending on their pick.